backend/requirements.txt
fastapi
uvicorn
requests
redis
pydantic
python-dotenv
httpx
